Fix service addition
