+++ src/Keda.Scaler.DurableTask.AzureStorage/Web/IServiceCollection.Extensions.cs
@@ -6,12 +6,9 @@ using Azure.Storage.Blobs;
 using Azure.Storage.Queues;
 using Keda.Scaler.DurableTask.AzureStorage.Accounts;
 using Keda.Scaler.DurableTask.AzureStorage.Common;
-using Keda.Scaler.DurableTask.AzureStorage.Security;
 using Keda.Scaler.DurableTask.AzureStorage.TaskHub;
-using Microsoft.AspNetCore.Authentication.Certificate;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
-using Microsoft.Extensions.Options;
 
 namespace Keda.Scaler.DurableTask.AzureStorage.Web;
 
@@ -26,40 +23,6 @@ internal static class IServiceCollectionExtensions
         services.TryAddSingleton<IOrchestrationAllocator, OptimalOrchestrationAllocator>();
         services.TryAddScoped<IProcessEnvironment>(p => new EnvironmentCache(ProcessEnvironment.Current));
         services.TryAddScoped<AzureStorageTaskHubBrowser>();
-
-        return services;
-    }
-
-    public static IServiceCollection AddTlsSupport(this IServiceCollection services)
-    {
-        ArgumentNullException.ThrowIfNull(services);
-
-        _ = services
-            .AddOptions<CertificateAuthenticationOptions>()
-            .BindConfiguration(TlsClientOptions.DefaultAuthenticationKey);
-
-        _ = services
-            .AddOptions<CertificateValidationCacheOptions>()
-            .BindConfiguration(TlsClientOptions.DefaultCachingKey);
-
-        _ = services
-            .AddAuthentication(CertificateAuthenticationDefaults.AuthenticationScheme)
-            .AddCertificate()
-            .AddCertificateCache();
-
-        _ = services
-            .AddOptions<TlsClientOptions>()
-            .BindConfiguration(TlsClientOptions.DefaultKey)
-            .ValidateDataAnnotations();
-
-        _ = services
-            .AddOptions<TlsServerOptions>()
-            .BindConfiguration(TlsServerOptions.DefaultKey)
-            .ValidateDataAnnotations();
-
-        services.TryAddSingleton<TlsConfigure>();
-        services.TryAddSingleton<IConfigureOptions<CertificateAuthenticationOptions>>(p => p.GetRequiredService<TlsConfigure>());
-        services.TryAddSingleton<IOptionsChangeTokenSource<CertificateAuthenticationOptions>>(p => p.GetRequiredService<TlsConfigure>());
 
         return services;
     }
